+++ tbr_deal_finder/__init__.py
@@ -1,5 +1,5 @@
 from pathlib import Path
 
-__VERSION__ = "0.4.2"
+__VERSION__ = "0.4.3"
 
 QUERY_PATH = Path(__file__).parent.joinpath("queries")